cleaned up & merged day detail modal together
