--- client/src/views/ContentCreator/LearningStandardCreator/CreateLessonDayEditor.less
@@ -1,18 +1,0 @@
-@import '../../../assets/style.less';
-
-#add-day-details {
-    input,
-    textarea {
-      display: block;
-      margin: auto;
-      background: none;
-      border: 1px solid;
-      border-color: #colors[secondary];
-      padding: 6px 6px;
-      width: 95%;
-      outline: none;
-      color: #colors[text-primary];
-      border-radius: 4px;
-      transition: 0.25s;
-    }
-}
